add hook for setting eventHandlers
