+++ src/components/animation/hooks.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { Props, WebGLContext } from './types';
-import { COLUMNS, createProgram, createShader, getDrawType, getFragmentShaderSource, getRandomColorVertices, getVertices, getVerticesCount, setupBuffers, setupUniforms } from './utils';
+import { COLUMNS, createProgram, createShader, getDrawType, getFragmentShaderSource, getRandomColorVertices, getVertices, getVerticesCount, setCanvasDimensions, setupBuffers, setupUniforms, updateBuffers, updateDynamicMode } from './utils';
 import { vertexShaderSource } from './shaders';
 
 const useWebGLContext = (
@@ -48,7 +48,6 @@ const useWebGLContext = (
     const isCircle = props.type === 'circle';
     const drawType = getDrawType(props.type, gl);
     let animationFrameId: number;
-    //let animationStartTime: DOMHighResTimeStamp | null = null;
     let elapsedTime = 0;
     verticesCount.current = getVerticesCount(props.type, vertices);
     const offsetTime = 0.8;
@@ -104,4 +103,107 @@ const useWebGLContext = (
   return isDeleted;
 };
 
-export { useWebGLContext };
+const useEvents = (props: Props, isDeleted: boolean, webGLContext: WebGLContext) => {
+  const { canvasRef, glRef, program, uniformsRef, animate, isDynamicMode, verticesCount, vBuffer, cBuffer, animationStartTime, currentNumVertices, isSmallScreen } = webGLContext;
+  useEffect(() => {
+    if (!canvasRef.current || !glRef.current || !program.current) return;
+    const canvas = canvasRef.current;
+    const gl = glRef.current;
+    const { u_animate, u_dynamic, u_width, u_resolution } = uniformsRef.current;
+
+
+    let newNumVertices = currentNumVertices.current;
+    const handleResize = () => {
+      gl.useProgram(program.current);
+      gl.uniform1f(u_width, window.innerWidth);
+
+      const isChanged = setCanvasDimensions(canvas, props.type);
+      if (isChanged) {
+        gl.viewport(0, 0, canvas.width, canvas.height);
+        gl.uniform2f(u_resolution, canvas.width, canvas.height);
+      }
+
+      // Dynamic mode can only be updated for circle type
+      if (props.type === 'circle') {
+        updateDynamicMode(gl, u_dynamic, isSmallScreen, isDynamicMode);
+        return;
+      }
+      // update buffer data if the number of vertices has changed
+      if (props.type !== 'shockwave') return; // Only for shockwave
+
+      if (window.innerWidth <= 600) {
+        newNumVertices = 50;
+      } else if (window.innerWidth > 600) {
+        newNumVertices = 100;
+      }
+
+      if (newNumVertices === currentNumVertices.current) return; // No change in vertices count
+      if (vBuffer.current === null || cBuffer.current === null || program.current === null) return; // Buffers not created
+      const newVertices = getVertices(props.type, newNumVertices);
+      updateBuffers(newVertices, gl, vBuffer.current, cBuffer.current, program.current);
+      verticesCount.current = getVerticesCount(props.type, newVertices); // vertices count represenst the number of data points to draw
+      currentNumVertices.current = newNumVertices;
+    };
+
+    const setAnimation = (value: boolean) => {
+      if (animate.current === value) return; // Already in the same state
+      gl.useProgram(program.current);
+      gl.uniform1i(u_animate, value ? 1 : 0);
+      animate.current = value;
+      animationStartTime.current = value ? performance.now() : null;
+    };
+    const isScrollendSupported = 'onscrollend' in window;
+    let isScrolling = false;
+    let timeout: NodeJS.Timeout;
+    const handleScroll = () => {
+      setAnimation(true);
+      isScrolling = true;
+      // Manually handle scrollend event if not supported by the browser
+      if (!isScrollendSupported) {
+        clearTimeout(timeout);
+        timeout = setTimeout(handleScrollEnd, 500);
+      }
+    };
+
+    const handleScrollEnd = () => {
+      setAnimation(isDynamicMode.current);
+      isScrolling = false;
+    };
+
+    const handleMouseEnter = () => {
+      if (isDynamicMode.current || isSmallScreen.current) return; // Already set or not allowed
+      gl.useProgram(program.current);
+      gl.uniform1i(u_dynamic, 1);
+      isDynamicMode.current = true;
+      setAnimation(true);
+    };
+
+    const handleMouseLeave = () => {
+      if (!isDynamicMode.current || isSmallScreen.current) return; // Already set or not allowed
+      gl.useProgram(program.current);
+      gl.uniform1i(u_dynamic, 0);
+      isDynamicMode.current = false;
+      setAnimation(isScrolling);
+    };
+
+    window.addEventListener('resize', handleResize);
+    if (props.type === 'circle') {
+      document.getElementById('about')?.addEventListener('mouseenter', handleMouseEnter);
+      document.getElementById('about')?.addEventListener('mouseleave', handleMouseLeave);
+      window.addEventListener('scroll', handleScroll);
+      window.addEventListener('scrollend', handleScrollEnd);
+    }
+    return () => {
+      window.removeEventListener('resize', handleResize);
+      if (props.type === 'circle') {
+        window.removeEventListener('scroll', handleScroll);
+        window.removeEventListener('scrollend', handleScrollEnd);
+        document.getElementById('about')?.removeEventListener('mouseenter', handleMouseEnter);
+        document.getElementById('about')?.removeEventListener('mouseleave', handleMouseLeave);
+        clearTimeout(timeout);
+      }
+    };
+  }, [isDeleted]);
+};
+
+export { useWebGLContext, useEvents };
